Add new optional parameters to Extract operation
Added the following optional parameters to fine-tune the behavior of
Extract operation:
- "include". This allows to specify which files should be extracted to
  the target directory.
- "search" and "replace". This allows to extract the contents of
  top-level directory or rename extracted files.
- "listFilePath". This allows to store a list of extracted files to the
  custom location. This list is used later to properly uninstall the
  component.

+++ src/libs/installer/extractarchiveoperation.cpp
@@ -29,6 +29,7 @@
 #include "extractarchiveoperation_p.h"
 
 #include "constants.h"
+#include "directoryguard.h"
 #include "globals.h"
 #include "fileguard.h"
 
@@ -74,12 +75,18 @@ ExtractArchiveOperation::ExtractArchiveOperation(PackageManagerCore *core)
 
 void ExtractArchiveOperation::backup()
 {
-    if (!checkArgumentCount(2))
+    if (!checkArgumentCount(2, 6))
         return;
 
     const QStringList args = arguments();
     const QString archivePath = args.at(0);
     const QString targetDir = args.at(1);
+    const QString include = args.value(2);
+    const QString search = args.value(3);
+    const QString replace = args.value(4);
+
+    QRegularExpression includeRegex(include);
+    QRegularExpression searchRegex(search);
 
     QScopedPointer<AbstractArchive> archive(ArchiveFactory::instance().create(archivePath));
     if (!archive) {
@@ -108,7 +115,15 @@ void ExtractArchiveOperation::backup()
     bool needsAdminRights = false;
 
     for (auto &entry : entries) {
-        const QString completeFilePath = targetDir + QDir::separator() + entry.path;
+        QString filePath = entry.path;
+        if (!include.isEmpty() && !filePath.contains(includeRegex))
+            continue;
+        if (!search.isEmpty()) {
+            filePath.replace(searchRegex, replace);
+            if (filePath.isEmpty())
+                continue;
+        }
+        const QString completeFilePath = targetDir + QDir::separator() + filePath;
         if (!entry.isDirectory) {
             // Ignore failed backups, existing files are overwritten when extracting.
             // Should the backups be used on rollback too, this may not be the
@@ -117,8 +132,8 @@ void ExtractArchiveOperation::backup()
         }
         if (!hasAdminRights && !canCreateSymLinks && entry.isSymbolicLink)
             needsAdminRights = true;
+        ++m_totalEntries;
     }
-    m_totalEntries = entries.size();
     if (needsAdminRights)
         setValue(QLatin1String("admin"), true);
 
@@ -128,19 +143,23 @@ void ExtractArchiveOperation::backup()
 
 bool ExtractArchiveOperation::performOperation()
 {
-    if (!checkArgumentCount(2))
+    if (!checkArgumentCount(2, 6))
         return false;
 
     const QStringList args = arguments();
     const QString archivePath = args.at(0);
     const QString targetDir = args.at(1);
+    const QString include = args.value(2);
+    const QString search = args.value(3);
+    const QString replace = args.value(4);
+    QString listFilePath = args.value(5);
 
     Receiver receiver;
     Callback callback;
 
     connect(&callback, &Callback::progressChanged, this, &ExtractArchiveOperation::progressChanged);
 
-    Worker *worker = new Worker(archivePath, targetDir, m_totalEntries, &callback);
+    Worker *worker = new Worker(archivePath, targetDir, include, search, replace, m_totalEntries, &callback);
     connect(worker, &Worker::finished, &receiver, &Receiver::workerFinished,
         Qt::QueuedConnection);
 
@@ -167,11 +186,10 @@ bool ExtractArchiveOperation::performOperation()
     // the dat file very slow to read and write. The .dat file is read into memory in startup,
     // writing the file names to a separate file we don't need to load all the file names into
     // memory as we need those only in uninstall. This will save a lot of memory.
-    // Parse a file and directorory structure using archivepath syntax
-    // installer://<component_name>/<filename>.7z Resulting structure is:
-    // -installerResources (dir)
-    //   -<component_name> (dir)
-    //    -<filename>.txt (file)
+    // If listFilePath is specified, the list of extracted files is stored there. Otherwise, it is
+    // stored in @TargetDir@/installerResources/<component_name>/<archive_name>.txt
+    // If listFilePath is "DATAFILE", the list of extracted files is written to the .dat file (the
+    // legacy behavior).
 
     QStringList files = callback.extractedFiles();
 
@@ -185,42 +203,73 @@ bool ExtractArchiveOperation::performOperation()
         installDir = core->value(scTargetDir);
         installerBaseBinary = core->toNativeSeparators(core->replaceVariables(core->installerBaseBinary()));
     }
-    const QString resourcesPath = installDir + QLatin1Char('/') + QLatin1String("installerResources");
 
-    QString fileDirectory = resourcesPath + QLatin1Char('/') + archivePath.section(QLatin1Char('/'), 1, 1,
-                            QString::SectionSkipEmpty) + QLatin1Char('/');
-    QString archiveFileName = archivePath.section(QLatin1Char('/'), 2, 2, QString::SectionSkipEmpty);
-    QFileInfo fileInfo2(archiveFileName);
-    QString suffix = fileInfo2.suffix();
-    archiveFileName.chop(suffix.length() + 1); // removes suffix (e.g. '.7z') from archive filename
-    QString fileName = archiveFileName + QLatin1String(".txt");
+    bool isDefaultListFile = listFilePath.isEmpty();
+    if (isDefaultListFile) {
+        // Parse a file and directory structure using archivepath syntax
+        // installer://<component_name>/<filename>.7z Resulting structure is:
+        // -installerResources (dir)
+        //   -<component_name> (dir)
+        //     -<filename>.txt (file)
+        const QString resourcesPath = installDir + QLatin1Char('/') + QLatin1String("installerResources");
 
-    QFileInfo targetDirectoryInfo(fileDirectory);
+        QString fileDirectory = resourcesPath + QLatin1Char('/') + archivePath.section(QLatin1Char('/'), 1, 1,
+                                QString::SectionSkipEmpty) + QLatin1Char('/');
+        QString archiveFileName = archivePath.section(QLatin1Char('/'), 2, 2, QString::SectionSkipEmpty);
+        QFileInfo fileInfo2(archiveFileName);
+        QString suffix = fileInfo2.suffix();
+        archiveFileName.chop(suffix.length() + 1); // removes suffix (e.g. '.7z') from archive filename
+        QString fileName = archiveFileName + QLatin1String(".txt");
 
-    QInstaller::createDirectoryWithParents(targetDirectoryInfo.absolutePath());
-    setDefaultFilePermissions(resourcesPath, DefaultFilePermissions::Executable);
-    setDefaultFilePermissions(targetDirectoryInfo.absolutePath(), DefaultFilePermissions::Executable);
+        QFileInfo targetDirectoryInfo(fileDirectory);
 
-    QFile file(targetDirectoryInfo.absolutePath() + QLatin1Char('/') + fileName);
-    if (file.open(QIODevice::WriteOnly)) {
-        setDefaultFilePermissions(file.fileName(), DefaultFilePermissions::NonExecutable);
-        QDataStream out (&file);
-        for (int i = 0; i < files.count(); ++i) {
-            if (!installerBaseBinary.isEmpty() && files[i].startsWith(installerBaseBinary)) {
-                // Do not write installerbase binary filename to extracted files. Installer binary
-                // is maintenance tool program, the binary is removed elsewhere
-                // when we do full uninstall.
-                files.clear();
-                break;
-            }
-            files[i] = replacePath(files.at(i), installDir, QLatin1String(scRelocatable));
+        QInstaller::createDirectoryWithParents(targetDirectoryInfo.absolutePath());
+        setDefaultFilePermissions(resourcesPath, DefaultFilePermissions::Executable);
+        setDefaultFilePermissions(targetDirectoryInfo.absolutePath(), DefaultFilePermissions::Executable);
+
+        listFilePath = targetDirectoryInfo.absolutePath() + QLatin1Char('/') + fileName;
+    } else if (listFilePath != scDataFile) {
+        const QString listDirPath = QFileInfo(listFilePath).absolutePath();
+        DirectoryGuard listDir(listDirPath);
+        try {
+            const QStringList createdDirs = listDir.tryCreate();
+            for (const QString &dir : createdDirs)
+                setDefaultFilePermissions(dir, DefaultFilePermissions::Executable);
+        } catch (const QInstaller::Error &error) {
+            qCWarning(QInstaller::lcInstallerInstallLog) << "Cannot create directory" << listDirPath << ":" << error.message();
         }
-        if (!files.isEmpty())
-            out << files;
-        setValue(QLatin1String("files"), file.fileName());
-        file.close();
+        listDir.release();
+    }
+
+    for (int i = 0; i < files.count(); ++i) {
+        if (!installerBaseBinary.isEmpty() && files[i].startsWith(installerBaseBinary)) {
+            // Do not write installerbase binary filename to extracted files. Installer binary
+            // is maintenance tool program, the binary is removed elsewhere
+            // when we do full uninstall.
+            files.clear();
+            break;
+        }
+    }
+
+    if (listFilePath != scDataFile) {
+        QFile file(listFilePath);
+        if (file.open(QIODevice::WriteOnly)) {
+            setDefaultFilePermissions(file.fileName(), DefaultFilePermissions::NonExecutable);
+            QDataStream out(&file);
+            for (int i = 0; i < files.count(); ++i)
+                files[i] = replacePath(files.at(i), installDir, QLatin1String(scRelocatable));
+            if (!files.isEmpty())
+                out << files;
+            setValue(QLatin1String("files"), file.fileName());
+            file.close();
+        } else {
+            qCWarning(QInstaller::lcInstallerInstallLog) << "Cannot open file for writing" << file.fileName() << ":" << file.errorString();
+        }
     } else {
-        qCWarning(QInstaller::lcInstallerInstallLog) << "Cannot open file for writing " << file.fileName() << ":" << file.errorString();
+        // Store the list of extracted files in maintenance tool .dat file (the legacy behavior).
+        for (int i = 0; i < files.count(); ++i)
+            files[i] = QDir::cleanPath(files.at(i));
+        setValue(QLatin1String("files"), files);
     }
 
     // TODO: Use backups for rollback, too? Doesn't work for uninstallation though.
@@ -239,7 +288,7 @@ bool ExtractArchiveOperation::performOperation()
 
 bool ExtractArchiveOperation::undoOperation()
 {
-    Q_ASSERT(arguments().count() == 2);
+    Q_ASSERT(arguments().count() >= 2 && arguments().count() <= 6);
 
     // For backward compatibility, check if "files" can be converted to QStringList.
     // If yes, files are listed in .dat instead of in a separate file.
@@ -251,16 +300,18 @@ bool ExtractArchiveOperation::undoOperation()
     if (useStringListType) {
         files = value(QLatin1String("files")).toStringList();
     } else {
-        if (!readDataFileContents(targetDir, &files))
+        if (!readListFileContents(targetDir, &files))
             return false;
     }
     if (!files.isEmpty())
         startUndoProcess(files);
     if (!useStringListType)
-        deleteDataFile(m_relocatedDataFileName);
+        deleteListFile(m_relocatedListFileName);
 
     // Remove the installerResources directory if it is empty.
-    QDir(targetDir).rmdir(QLatin1String("installerResources"));
+    bool isDefaultListFile = arguments().value(5).isEmpty();
+    if (isDefaultListFile)
+        QDir(targetDir).rmdir(QLatin1String("installerResources"));
 
     return true;
 }
@@ -283,20 +334,25 @@ void ExtractArchiveOperation::startUndoProcess(const QStringList &files)
     thread->deleteLater();
 }
 
-void ExtractArchiveOperation::deleteDataFile(const QString &fileName)
+void ExtractArchiveOperation::deleteListFile(const QString &fileName)
 {
     if (fileName.isEmpty()) {
-        qCWarning(QInstaller::lcInstallerInstallLog) << Q_FUNC_INFO << "data file name cannot be empty.";
+        qCWarning(QInstaller::lcInstallerInstallLog) << Q_FUNC_INFO << "list file name cannot be empty.";
         return;
     }
     QFile file(fileName);
     QFileInfo fileInfo(file);
     if (file.remove()) {
         QDir directory(fileInfo.absoluteDir());
-        if (directory.exists() && directory.isEmpty())
-            directory.rmdir(directory.path());
+        if (directory.exists() && directory.isEmpty()) {
+            bool isDefaultListFile = arguments().value(5).isEmpty();
+            if (isDefaultListFile)
+                directory.rmdir(directory.path());
+            else
+                directory.rmpath(directory.path());
+        }
     } else {
-        qCWarning(QInstaller::lcInstallerInstallLog) << "Cannot remove data file" << file.fileName();
+        qCWarning(QInstaller::lcInstallerInstallLog) << "Cannot remove list file" << file.fileName();
     }
 }
 
@@ -336,7 +392,7 @@ bool ExtractArchiveOperation::testOperation()
 
 quint64 ExtractArchiveOperation::sizeHint()
 {
-    if (!checkArgumentCount(2))
+    if (!checkArgumentCount(2, 6))
         return UpdateOperation::sizeHint();
 
     if (hasValue(QLatin1String("sizeHint")))
@@ -352,14 +408,14 @@ quint64 ExtractArchiveOperation::sizeHint()
     return compressedSize;
 }
 
-bool ExtractArchiveOperation::readDataFileContents(QString &targetDir, QStringList *resultList)
+bool ExtractArchiveOperation::readListFileContents(QString &targetDir, QStringList *resultList)
 {
     const QString filePath = value(QLatin1String("files")).toString();
     // Does not change target on non macOS platforms.
     if (QInstaller::isInBundle(targetDir, &targetDir))
         targetDir = QDir::cleanPath(targetDir + QLatin1String("/.."));
-    m_relocatedDataFileName = replacePath(filePath, QLatin1String(scRelocatable), targetDir);
-    QFile file(m_relocatedDataFileName);
+    m_relocatedListFileName = replacePath(filePath, QLatin1String(scRelocatable), targetDir);
+    QFile file(m_relocatedListFileName);
 
     if (file.open(QIODevice::ReadOnly)) {
         QDataStream in(&file);
